Split config and client into separate files
- change some names and text to be more consistent within shuttle
- add some more client methods
- remove unnecessary args in methods
